pyproject.toml
[tool.poetry]
name = "aicapi_yw3760"
version = "0.1.0"
description = "This is a python package of an API CLIENT to retrieve the data of collections in the Atr Institute of Chicago."
authors = ["yw3760"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.5"
Pillow = "^8.4.0"
ipython = "^7.30.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
myst-nb = "^0.13.1"
sphinx-autoapi = "^1.8.4"
sphinx-rtd-theme = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
